+++ src/physics.rs
@@ -1,6 +1,11 @@
 use crate::sprite::Sprite;
 use bevy::prelude::*;
-use std::{collections::HashSet, hash::Hash};
+use serde::{Deserialize, Serialize};
+use std::{
+    collections::HashSet,
+    f32::consts::{PI, TAU},
+    hash::Hash,
+};
 
 pub struct PhysicsPlugin;
 
@@ -11,6 +16,10 @@ impl Plugin for PhysicsPlugin {
     }
 }
 
+/// This is the struct that is generated when a collision occurs. Collisions only occur between two
+/// [Sprite]s which:
+/// - have colliders (you can use the `collider_creator` example to create your own colliders)
+/// - have their `collision` flags set to `true`.
 #[derive(Debug, Clone)]
 pub struct CollisionEvent {
     pub state: CollisionState,
@@ -78,6 +87,7 @@ impl Hash for CollisionPair {
     }
 }
 
+// system - detect collisions and generate the collision events
 fn collision_detection(
     mut existing_collisions: Local<HashSet<CollisionPair>>,
     mut collision_events: EventWriter<CollisionEvent>,
@@ -126,7 +136,9 @@ fn collision_detection(
     }
 }
 
-#[derive(Clone, Debug)]
+/// Represents the collider (or lack thereof) of a sprite. Two sprites need to have colliders AND
+/// have their `Sprite.collision` fields set to `true` to generate collision events.
+#[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
 pub enum Collider {
     NoCollider,
     Poly(Vec<Vec2>),
@@ -139,6 +151,7 @@ impl Default for Collider {
 }
 
 impl Collider {
+    /// Generate a rectangular collider based on top-left and bottom-right points
     pub fn rect<T: Into<Vec2>>(topleft: T, bottomright: T) -> Self {
         let topleft = topleft.into();
         let bottomright = bottomright.into();
@@ -149,25 +162,103 @@ impl Collider {
             Vec2::new(topleft.x, bottomright.y),
         ])
     }
+    /// Convert a slice of Vec2's into a polygon collider. This is helpful if you want to hard-code
+    /// colliders in your code as arrays or vectors of Vec2.
     pub fn poly<T: Into<Vec2> + Copy>(points: &[T]) -> Self {
         Self::Poly(points.iter().map(|&x| x.into()).collect())
     }
+    /// Generate a polygon circle approximation with a specified amount of vertices
     pub fn circle_custom(radius: f32, vertices: usize) -> Self {
         let mut points = vec![];
-        for x in 0..=vertices {
-            let inner = 2.0 * std::f64::consts::PI / vertices as f64 * x as f64;
-            points.push(Vec2::new(
-                inner.cos() as f32 * radius,
-                inner.sin() as f32 * radius,
-            ));
+        for x in 0..vertices {
+            let inner = std::f64::consts::TAU / vertices as f64 * x as f64;
+            let mut inner_x = inner.cos() as f32 * radius;
+            let mut inner_y = inner.sin() as f32 * radius;
+            // Clamp near-zero values to zero when producing RON files: (-0.0000000000000044087286)
+            if (inner_x > -0.000001) && (inner_x < 0.000001) {
+                inner_x = 0.0;
+            }
+            if (inner_y > -0.000001) && (inner_y < 0.000001) {
+                inner_y = 0.0;
+            }
+            points.push(Vec2::new(inner_x, inner_y));
         }
         Self::Poly(points)
     }
+    /// Generate a 16-vertex polygon circle approximation. 16 was chosen as the default as it works
+    /// quite well with the circular sprites in the asset pack.
     pub fn circle(radius: f32) -> Self {
         Self::circle_custom(radius, 16)
     }
+    /// Whether or not the collider is a `Collider::Poly`.
     pub fn is_poly(&self) -> bool {
         matches!(self, Self::Poly(_))
+    }
+    /// Whether the points in the collider represent a convex polygon (not concave or complex).
+    /// This is important, because Rusty Engine's collision detection doesn't work correctly unless
+    /// colliders are convex polygons.
+    ///
+    /// Implementation based on Rory Daulton's answer on https://stackoverflow.com/questions/471962/how-do-i-efficiently-determine-if-a-polygon-is-convex-non-convex-or-complex?answertab=votes#tab-top
+    pub fn is_convex(&self) -> bool {
+        if let Collider::Poly(points) = self {
+            let length = points.len();
+            if length < 3 {
+                return false; // empty sets, points and lines are not convex polygons
+            }
+            // the source algorithm deals with individual x's and y's and the combined points in
+            // disjoint ways, so we need to follow the pattern unless we want to modify the
+            // algorithm itself.
+            let mut old_x = points[length - 2].x;
+            let mut old_y = points[length - 2].y;
+            let mut new_x = points[length - 1].x;
+            let mut new_y = points[length - 1].y;
+            let mut new_direction = (new_y - old_y).atan2(new_x - old_x);
+            let mut angle_sum = 0.0;
+            let mut old_direction;
+            let mut orientation = 0.0;
+            for (idx, newpoint) in points.iter().enumerate() {
+                // The fact that new_x and new_y are re-used at the top of the loop with the
+                // expectation that they have the last loop's values is why we can't use the
+                // newpoint loop variable directly. Messy. :-/
+                old_x = new_x;
+                old_y = new_y;
+                old_direction = new_direction;
+                new_x = newpoint.x;
+                new_y = newpoint.y;
+                new_direction = (new_y - old_y).atan2(new_x - old_x);
+                if (old_x == new_x) && (old_y == new_y) {
+                    return false; // repeated consecutive points
+                }
+                // Calculate & check the normalized deriction-change angle
+                let mut angle = new_direction - old_direction;
+                if angle <= -PI {
+                    angle += TAU; // make it in half-open interval (-Pi, Pi]
+                } else if angle > PI {
+                    angle -= TAU;
+                }
+                if idx == 0 {
+                    // if first time through loop, initialize orientation
+                    if angle == 0.0 {
+                        return false; // the source algorithm doesn't explain this one
+                    }
+                    if angle > 0.0 {
+                        orientation = 1.0;
+                    } else {
+                        orientation = -1.0;
+                    }
+                } else if orientation * angle <= 0.0 {
+                    // not both positive or both negative
+                    return false;
+                }
+
+                // Accumulate the direction-change angle
+                angle_sum += angle;
+            }
+            // Check that the total number of full turns is plus-or-minus 1
+            let full_turns = (angle_sum / TAU).abs();
+            return (full_turns > 0.9999) && (full_turns < 1.0001);
+        }
+        false
     }
     fn rotated(&self, rotation: f32) -> Vec<Vec2> {
         let mut rotated_points = Vec::new();
@@ -183,7 +274,9 @@ impl Collider {
         }
         rotated_points
     }
-    fn relative_to(&self, sprite: &Sprite) -> Vec<Vec2> {
+    #[doc(hidden)]
+    // Used internally to scale colliders to match a sprite's current transform
+    pub fn relative_to(&self, sprite: &Sprite) -> Vec<Vec2> {
         self.rotated(sprite.rotation)
             .iter()
             .map(|&v| v * sprite.scale + sprite.translation) // scale & translation